api/requirements.txt
flask==3.0.0
gunicorn==21.2.0
google-cloud-dataproc==5.11.0
google-cloud-storage==2.18.2
protobuf==4.25.5
python-dotenv==1.0.1
